deployment/ray_serve/requirements-k8s.txt
ray[serve]>=2.9.0,<2.40.0
fastapi>=0.110.0
uvicorn>=0.29.0
requests>=2.31.0
psutil>=5.9.0
numpy>=1.24.0,<2.0.0
transformers>=4.35.0,<5.0.0
huggingface-hub>=0.20.0,<1.0.0
pydantic>=2.0.0

# Observability
prometheus-client>=0.20.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-http>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-requests>=0.45b0
